fix: move layout to features/layout folder
