+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import { type NextPage } from "next"
-import { Layout } from "~/components/Layout"
+import { Layout } from "~/features/layout"
 
 const Home: NextPage = () => {
 	return <Layout></Layout>
